+++ org.osgi.service.zigbee/src/org/osgi/service/zigbee/types/ZigBeeBitmap64.java
@@ -17,7 +17,6 @@
 package org.osgi.service.zigbee.types;
 
 import java.io.IOException;
-import java.math.BigInteger;
 import org.osgi.service.zigbee.ZigBeeDataInput;
 import org.osgi.service.zigbee.ZigBeeDataOutput;
 import org.osgi.service.zigbee.ZigBeeDataTypes;
@@ -53,7 +52,7 @@ public class ZigBeeBitmap64
 	}
 
 	public Class getJavaDataType() {
-		return BigInteger.class;
+		return Long.class;
 	}
 
 	public short getId() {
