+++ org.osgi.service.zigbee/src/org/osgi/service/zigbee/types/ZigBeeBitmap8.java
@@ -52,7 +52,7 @@ public class ZigBeeBitmap8
 	}
 
 	public Class getJavaDataType() {
-		return Short.class;
+		return Byte.class;
 	}
 
 	public short getId() {
